refactor(DeploymentsPanel): fix imports
