+++ lib/manager/containers/ActiveDeploymentViewer.js
@@ -13,7 +13,7 @@ import {
   updateDeployment,
   updateVersionForFeedSource
 } from '../actions/deployments'
-import DeploymentViewer from '../components/DeploymentViewer'
+import DeploymentViewer from '../components/deployment/DeploymentViewer'
 
 import type {Deployment, Feed, Project} from '../../types'
 import type {AppState} from '../../types/reducers'
