+++ lib/manager/containers/DeploymentsPanel.js
@@ -11,7 +11,7 @@ import {
   deleteDeployment,
   updateDeployment
 } from '../actions/deployments'
-import DeploymentsPanel from '../components/DeploymentsPanel'
+import DeploymentsPanel from '../components/deployment/DeploymentsPanel'
 
 import type {Project} from '../../types'
 import type {AppState} from '../../types/reducers'
